Make SimpleSOL tests work with MPI.

+++ test/integration/solvers/solver_test_utils.cpp
@@ -1,4 +1,5 @@
 #include "solver_test_utils.h"
+#include <mpi.h>
 
 // ======================================= Helper functions =======================================
 std::filesystem::path get_common_test_resources_dir(std::string solver_name) {
@@ -14,6 +15,18 @@ std::filesystem::path get_test_resources_dir(std::string solver_name, std::strin
 
 std::filesystem::path get_test_run_dir(std::string solver_name, std::string test_name) {
   return std::filesystem::temp_directory_path() / "neso-tests/solvers" / solver_name / test_name;
+}
+
+int get_rank() {
+  int rank;
+  MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+  return rank;
+}
+
+bool is_root() {
+  int size;
+  MPI_Comm_size(MPI_COMM_WORLD, &size);
+  return (size < 2 || get_rank() == 0);
 }
 
 // Get solver name from test_suite_name, removing "Test" suffix if necessary
@@ -52,29 +65,36 @@ std::vector<std::string> NektarSolverTest::get_default_args() {
 }
 
 void NektarSolverTest::make_test_run_dir() {
-  // Remove any previous run dir
-  std::filesystem::remove_all(m_test_run_dir);
-  // Create dir
-  std::filesystem::create_directories(m_test_run_dir.parent_path());
+  if (is_root()) {
+    // Remove any previous run dir
+    std::filesystem::remove_all(m_test_run_dir);
+    // Create dir
+    std::filesystem::create_directories(m_test_run_dir.parent_path());
 
-  // Copy in common resources
-  if (std::filesystem::exists(m_common_test_res_dir)) {
-    std::filesystem::copy(m_common_test_res_dir, m_test_run_dir,
-                          std::filesystem::copy_options::recursive |
-                              std::filesystem::copy_options::overwrite_existing);
-  } else {
-    std::cout << "Skipping copy of common resources; no directory at " << m_common_test_res_dir
-              << std::endl;
+    // Copy in common resources
+    if (std::filesystem::exists(m_common_test_res_dir)) {
+      std::filesystem::copy(m_common_test_res_dir, m_test_run_dir,
+                            std::filesystem::copy_options::recursive |
+                                std::filesystem::copy_options::overwrite_existing);
+    } else {
+      std::cout << "Skipping copy of common resources; no directory at " << m_common_test_res_dir
+                << std::endl;
+    }
+
+    // Copy in test-specific resources
+    if (std::filesystem::exists(m_test_res_dir)) {
+      std::filesystem::copy(m_test_res_dir, m_test_run_dir,
+                            std::filesystem::copy_options::recursive |
+                                std::filesystem::copy_options::overwrite_existing);
+    } else {
+      std::cout << "Skipping copy of test-specific resources; no directory at " << m_test_res_dir
+                << std::endl;
+    }
   }
-
-  // Copy in test-specific resources
-  if (std::filesystem::exists(m_test_res_dir)) {
-    std::filesystem::copy(m_test_res_dir, m_test_run_dir,
-                          std::filesystem::copy_options::recursive |
-                              std::filesystem::copy_options::overwrite_existing);
-  } else {
-    std::cout << "Skipping copy of test-specific resources; no directory at " << m_test_res_dir
-              << std::endl;
+  // Other tasks wait for the dir to be created before continuuing
+  int state_at_barrier = MPI_Barrier(MPI_COMM_WORLD);
+  if (MPI_SUCCESS != state_at_barrier) {
+    std::cout << "MPI_Barrier failed" << std::endl;
   }
 }
 
@@ -93,9 +113,11 @@ void NektarSolverTest::redirect_output_to_file() {
   m_orig_stdout_buf = std::cout.rdbuf();
   m_orig_stderr_buf = std::cerr.rdbuf();
   // Redirect
-  m_out_strm.open(m_test_run_dir / "stdout.txt");
+  std::string stdout_fname = "stdout." + std::to_string(get_rank()) + ".txt";
+  std::string stderr_fname = "stderr." + std::to_string(get_rank()) + ".txt";
+  m_out_strm.open(m_test_run_dir / stdout_fname);
   std::cout.rdbuf(m_out_strm.rdbuf());
-  m_err_strm.open(m_test_run_dir / "stderr.txt");
+  m_err_strm.open(m_test_run_dir / stderr_fname);
   std::cerr.rdbuf(m_err_strm.rdbuf());
 }
